poll data for three-stage conversion process
